transfer connection responsibility to Service layer, cnx pool tuned, refactor

+++ src/main/java/com/excilys/formation/computerdatabase/service/ComputerService.java
@@ -1,12 +1,18 @@
 package com.excilys.formation.computerdatabase.service;
 
+import java.sql.Connection;
+import java.sql.SQLException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.excilys.formation.computerdatabase.exceptions.DAOException;
 import com.excilys.formation.computerdatabase.model.Computer;
 import com.excilys.formation.computerdatabase.model.Page;
 import com.excilys.formation.computerdatabase.model.SelectOptions;
 import com.excilys.formation.computerdatabase.persistence.ComputerDAO;
+import com.excilys.formation.computerdatabase.persistence.ComputerDAOable;
+import com.excilys.formation.computerdatabase.persistence.ConnectionFactory;
 import com.excilys.formation.computerdatabase.persistence.Crudable;
 
 /**
@@ -20,13 +26,20 @@ import com.excilys.formation.computerdatabase.persistence.Crudable;
  *
  */
 public class ComputerService implements Servable<Computer> {
+  
+  public static final ThreadLocal<Connection> localConnection = new ThreadLocal<>();
+  
   private static final Logger LOGGER =
       LoggerFactory.getLogger("com.excilys.formation.computerdatabase");
 
-  private final Crudable<Computer> cdao;
+  private final ConnectionFactory connectionFactory;
+
+  private final ComputerDAOable cdao;
 
   private ComputerService() {
     this.cdao = ComputerDAO.INSTANCE;
+    this.connectionFactory = ConnectionFactory.getInstance();
+//    this.localConnection = new ThreadLocal<>();
   }
 
   /**
@@ -46,8 +59,17 @@ public class ComputerService implements Servable<Computer> {
   @Override
   public int count() {
     LOGGER.info("count " + this.getClass());
+    
+    this.initConnection();
 
-    int count = this.cdao.count();
+    int count = 0;
+    try {
+      count = this.cdao.count();
+    } catch (DAOException e) {
+      LOGGER.error(e.getMessage());
+    } finally {
+      this.closeConnection();
+    }
 
     return count;
   }
@@ -55,8 +77,17 @@ public class ComputerService implements Servable<Computer> {
   @Override
   public int count(SelectOptions options) {
     LOGGER.info("count with options" + this.getClass());
+    
+    this.initConnection();
 
-    int count = this.cdao.count(options);
+    int count = 0;
+    try {
+      count = this.cdao.count(options);
+    } catch (DAOException e) {
+      LOGGER.error(e.getMessage());
+    } finally {
+      this.closeConnection();
+    }
 
     return count;
   }
@@ -65,8 +96,16 @@ public class ComputerService implements Servable<Computer> {
   public Page<Computer> get(SelectOptions options) {
     LOGGER.info("get options " + this.getClass());
 
-    Page<Computer> computerPage =
-        new Page<>(options.getOffset(), this.cdao.find(options), this.cdao.count());
+    this.initConnection();
+    
+    Page<Computer> computerPage = null;
+    try {
+      computerPage = new Page<>(options.getOffset(), this.cdao.find(options), this.cdao.count());
+    } catch (DAOException e) {
+      LOGGER.error(e.getMessage());
+    } finally {
+      this.closeConnection();
+    }
 
     return computerPage;
   }
@@ -74,29 +113,66 @@ public class ComputerService implements Servable<Computer> {
   @Override
   public Computer get(Long id) {
     LOGGER.info("get id " + this.getClass());
-    return this.cdao.find(id);
+    
+    this.initConnection();
+    
+    Computer result = null;
+    try {
+      result = this.cdao.find(id);
+    } catch (DAOException e) {
+      LOGGER.error(e.getMessage());
+    } finally {
+      this.closeConnection();
+    }
+    
+    return result;
   }
 
   @Override
   public Long create(Computer computer) {
     LOGGER.info("create " + this.getClass());
     LOGGER.debug(computer.toString());
+    
+    this.initConnection();
+    
     if (computer.getName() == null || computer.getName().isEmpty()) {
       LOGGER.error("ERROR Insert : Could not create an unnamed computer !");
       return null;
     }
-    if (computer.getCompany().getId() == null) {
-      LOGGER.error("ERROR Insert : Could not create a computer without it's company !!");
-      return null;
+    
+    Long result = null;
+    try {
+      result = this.cdao.create(computer);
+    } catch (DAOException e) {
+      LOGGER.error(e.getMessage());
+    } finally {
+      this.closeConnection();
     }
-    return this.cdao.create(computer);
+    
+    return result;
   }
 
   @Override
   public boolean delete(Long id) {
     LOGGER.info("delete " + this.getClass());
+    if(id == null){
+      return false;
+    }
+    
     LOGGER.debug("Delete Computer with id " + id);
-    return this.cdao.delete(id);
+
+    this.initConnection();
+    
+    boolean result = false;
+    try {
+      result = this.cdao.delete(id);
+    } catch (DAOException e) {
+      LOGGER.error("Delete : failed, no rows affected");
+    } finally {
+      this.closeConnection();
+    }
+
+    return result;
   }
 
   @Override
@@ -113,7 +189,45 @@ public class ComputerService implements Servable<Computer> {
       LOGGER.error("ERROR Update : Could not update to a computer without it's company !!");
       return null;
     }
+    
+    this.initConnection();
+    
+    Computer result = null;
 
-    return this.cdao.update(computer);
+    try {
+      result = this.cdao.update(computer);
+    } catch (DAOException e) {
+      LOGGER.error(e.getMessage());
+    } finally {
+      this.closeConnection();
+    }
+    
+    return result;
+  }
+  
+  /**
+   * Those to functions areNot in the interface.
+   * It would be dangerous to expose that.
+   */
+  
+  /**
+   * Sets the connection object in the LocalThread and returns it.
+   * @return
+   */
+  private Connection initConnection() {
+    Connection connection = this.connectionFactory.getConnection();
+
+    ComputerService.localConnection.set(connection);
+
+    return connection;
+  }
+  /**
+   * Closes the connection.
+   * And cleans the LocalThread
+   * @param connection
+   */
+  private void closeConnection(){
+    this.connectionFactory.closeConnection(ComputerService.localConnection.get());
+    ComputerService.localConnection.remove();
   }
 }
